go-controller: Move some functions to pkg.
Some of the functions in nics-to-bridge will be
reused in an upcoming commit.
